Add files via upload
This code goes into the nano keyboard implant.  The key packet formatting will be moved to the Xbee eventually.  Insomnia mode is added here on its own BLE Characteristic.  There is the beginning of a mode select that isn't completely flushed out.  It will be used later with the Xbee C2.