Add files via upload
This file is to be loaded into PyCharm with the DigiMesh 2.4 firmware and flashed over to the XBee3.  It uses Python 3.8.  The code here will work with the BLECentral code to receive keystroke packets, map them to keys presses and relay them out to the C2.  This code will also relay commands from the C2 to the BLECentral device (and ultimately to the BLEPeripheral).